refactor(presentation/home): moving spies, mocks for helpers folder
