+++ tests/presentation/home/useHome.spec.tsx
@@ -1,90 +1,14 @@
 import { Place } from 'src/presentation/components/cardList';
 import { renderHook, waitFor } from '@testing-library/react-native';
-import { faker } from '@faker-js/faker';
-import { ListPlaces, ListRecommendations } from '~/domain/useCases';
-import { PlaceModel, RecommendationModel } from '~/domain/models';
+import { RecommendationModel } from '~/domain/models';
 import useHome from '../../../src/presentation/home/useHome';
-import placeListFactory from '../helpers/placeListFactory';
 import { LocationServiceSpy } from '../helpers/locationServiceSpy';
+import { ListRecommendationsSpy } from '../helpers/listRecommendationsSpy';
+import { ListPlacesSpy } from '../helpers/listPlacesSpy';
+import { recommendationModelFactory } from '../helpers/recommendationModelFactory';
+import placeListFactory from '../helpers/placeListFactory';
 
 jest.useFakeTimers();
-
-const recommendationModelFake = (): RecommendationModel => {
-  return {
-    location: faker.address.secondaryAddress(),
-    imageUrl: faker.image.city(),
-    title: faker.name.jobTitle(),
-    rating: faker.datatype
-      .number({ min: 1, max: 10, precision: 0.1 })
-      .toString(),
-    myDistanceOfLocal: faker.datatype.number().toString(),
-    id: faker.datatype.number(),
-  };
-};
-
-class ListRecommendationsSpy implements ListRecommendations {
-  listCalled = 0;
-  timeout = 0;
-
-  constructor(
-    readonly recommendations: Array<RecommendationModel> = [
-      recommendationModelFake(),
-    ],
-  ) {}
-  async list(): Promise<RecommendationModel[]> {
-    let recommendations: Array<RecommendationModel> = [];
-    const complete = () => {
-      this.listCalled += 1;
-      recommendations = this.recommendations;
-    };
-
-    if (this.timeout > 0) {
-      setTimeout(() => {
-        complete();
-      }, this.timeout);
-    } else {
-      complete();
-    }
-
-    return recommendations;
-  }
-
-  addTimeout(timeout: number) {
-    this.timeout = timeout;
-  }
-}
-
-export class ListPlacesSpy implements ListPlaces {
-  listCalled = 0;
-  timeout = 0;
-  constructor(readonly places: Array<PlaceModel> = placeListFactory(5)) {}
-  async list(
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    location: { long: string; lat: string },
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    nextPageToken?: string | undefined,
-  ): Promise<PlaceModel[]> {
-    let places: Array<PlaceModel> = [];
-    const complete = () => {
-      this.listCalled += 1;
-      places = this.places;
-    };
-
-    if (this.timeout > 0) {
-      setTimeout(() => {
-        complete();
-      }, this.timeout);
-    } else {
-      complete();
-    }
-
-    return places;
-  }
-
-  addTimeout(timeout: number) {
-    this.timeout = timeout;
-  }
-}
 
 describe('Presentation: useHome', () => {
   test('should call navigate function correctly when call onSeeAll function with Discover param', async () => {
@@ -104,7 +28,7 @@ describe('Presentation: useHome', () => {
   });
 
   test('should get the recommendations through of ListRecommendations when initialize', async () => {
-    const recommendations = [recommendationModelFake()];
+    const recommendations = [recommendationModelFactory()];
     const listRecommendations = new ListRecommendationsSpy(recommendations);
     const {
       sut: { result },
@@ -201,7 +125,7 @@ describe('Presentation: useHome', () => {
     test('should the error returning false when recommendations and place list are empty and finding is true', async () => {
       const listPlaces = new ListPlacesSpy(placeListFactory(5));
       const listRecommendations = new ListRecommendationsSpy([
-        recommendationModelFake(),
+        recommendationModelFactory(),
       ]);
       listPlaces.addTimeout(1000);
       listRecommendations.addTimeout(1000);
@@ -247,7 +171,7 @@ describe('Presentation: useHome', () => {
     test('should the finding returning true correctly when it is finding the list recommendation and place', async () => {
       const listPlaces = new ListPlacesSpy(placeListFactory(5));
       const listRecommendations = new ListRecommendationsSpy([
-        recommendationModelFake(),
+        recommendationModelFactory(),
       ]);
       listPlaces.addTimeout(1000);
       listRecommendations.addTimeout(1000);
@@ -269,7 +193,7 @@ describe('Presentation: useHome', () => {
     test('should set finding to true and then back to false when calling tryGetListAgain', async () => {
       const listPlaces = new ListPlacesSpy(placeListFactory(5));
       const listRecommendations = new ListRecommendationsSpy([
-        recommendationModelFake(),
+        recommendationModelFactory(),
       ]);
 
       const {
@@ -392,7 +316,9 @@ type SutProps = {
 
 const makeSut = ({
   listPlaces = new ListPlacesSpy(placeListFactory(5)),
-  listRecommendations = new ListRecommendationsSpy([recommendationModelFake()]),
+  listRecommendations = new ListRecommendationsSpy([
+    recommendationModelFactory(),
+  ]),
   locationService = new LocationServiceSpy(),
 }: SutProps) => {
   const navigate = jest.fn();
